Edit create invoice to create journal

+++ Central App/src/main/java/com/graduationProject/_thYear/Journal/dtos/request/CreateJournalRequest.java
@@ -26,7 +26,7 @@ public class CreateJournalRequest {
     @NotNull(message = "Currency value is required")
     private BigDecimal currencyValue;
 
-    private Byte parentType;
+    private Integer parentType;
     private Integer parentId;
 
     @Builder.Default
