+++ Central App/src/main/java/com/graduationProject/_thYear/Journal/models/JournalHeader.java
@@ -59,7 +59,7 @@ public class JournalHeader {
     private BigDecimal currencyValue;
 
     @Column(name = "parentType")
-    private Byte parentType;
+    private Integer parentType;
 
     @Convert(converter = JournalKindConverter.class)
     @Column(name = "kind")
